Implemented trigger dependent rQCD
Implemented different rQCD values for the 1p and 3p events. (This feature was
actually missing until now.)

Implemented different rQCD values for the total and passed estimations.

Updated list of pruned systematics.

Implemented uncorrelated variation of the total and the passed for RQCD_SYST
and RQCD_STAT. Switched to using correlated variation of the total and the
passed for all other systematic uncertainties.