Plotting and computation script improvements
Added protection against ZeroDivisionError in rQCD computation.
Added multiple options to control the ATLAS label and the stamp from the
execution script for both the generic and the efficiency plotting scripts.
Added an inclusive trigger efficiency computation and options to control
the exclusion of inclusive or prong-separated efficiencies.
Added new distributions, and fixed brackets around units.