Updated plotting and configuration for v12 ntuple
Added year dependent treatment of the trigger SF/uncertainties.
Added treatment of pileup systematics.
Added ttbar CR to regions.
Updated treatment of uncertainties, so that owls_mutau/uncertainties.py takes
an uncertainties configuration which belongs in the model. Default
configuration is for MC15B, but should be moved to MC15B model.
Minor improvements to plot.py and plot-syst-variation.py.

NOTE: Change to systematics treatment is not complete for mc15b.