Plotting aesthetics and more 2015 efficiencies.
Fixed some plotting aesthetics such as
* Background label
* Region label
* No ATLAS stamp by default
* Plot margins

Added estimation of systematic effect on the total weighted trigger efficiency
through a least squares regression. This estimate is more relevant to the
efficiency than that of the systematic effect on the background yields.

Added tau50, tau80, tau125, and tau160 measurements also for 2015 data.